requirements.txt
Django==5.0.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
# django-cryptography==1.1  # Incompatible avec Django 5.0
psycopg2-binary==2.9.7
celery==5.3.1
redis==4.6.0
python-decouple==3.8
requests==2.31.0
Pillow==10.0.0
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.5.0
drf-yasg==1.21.7
